Fix test issues due to DateTimeUtils changes
The DateTimeUtils object has been adjusted in 4.5.2 to enforce Instant
timestamps without a timezone and up to ms. As the EventMessage itself
can still contain an entire Instant whilst the converters convert
through the utils class, test issues pop up. The expected timestamp for
the tests should be adjusted to expect a different result.

#97